requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.2
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.0
appdirs==1.4.4
argumentparser==1.2.1
asttokens==3.0.0
async-timeout==4.0.2
attrs==25.3.0
audioread==3.0.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.13.4
binary==1.0.0
bitarray==2.7.6
bleach==6.2.0
blinker==1.6.2
blis==0.7.11
bottle==0.12.23
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
confection==0.1.3
contourpy==1.1.0
crashtest==0.4.1
cryptography==42.0.8
cycler==0.11.0
cymem==2.0.8
Cython==3.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docopt==0.6.2
docx2txt==0.8
dulwich==0.21.5
easygui==0.98.3
easyocr==1.7.1
editdistpy==0.1.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.6.1/en_core_web_trf-3.6.1-py3-none-any.whl#sha256=f9a1f0ae83954c49ebf92c0f7043ea02de51db59bf6cd3ed849d46928612c806
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==2.2.0
fairseq==0.12.2
faiss-cpu==1.7.4
fastapi==0.100.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
fire==0.6.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gingerit==0.0.0.1
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gradio==3.37.0
gradio_client==0.2.10
greenlet==2.0.2
gTTS==2.3.2
h11==0.12.0
html5lib==1.1
httpcore==0.13.7
httplib2==0.22.0
httpx==1.0.0b0
huggingface-hub==0.24.6
hydra-core==1.0.7
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.6.0
importlib-metadata==6.8.0
ingress==0.3.2
iniconfig==2.0.0
install==1.3.5
installer==0.7.0
ipython==8.12.3
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.6.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyboard==0.13.5
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
language_tool_python==2.8.1
lazy_loader==0.3
libaudioverse==0.9.1
librosa==0.9.1
linkify-it-py==2.0.2
llvmlite==0.40.1
lockfile==0.12.2
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.1.3
more-itertools==9.1.0
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.10
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.57.1
numpy==2.2.6
omegaconf==2.0.6
openai==0.28.1
openai-whisper==20240930
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
optional==0.0.1
orjson==3.9.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathy==0.10.2
patool==1.12
pdf2docx==0.5.8
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
pkginfo==1.9.6
platformdirs==3.9.1
playsound==1.3.0
pluggy==1.5.0
plyer==2.1.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
pooch==1.6.0
portalocker==2.7.0
praat-parselmouth==0.4.3
preshed==3.0.9
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
PubChemPy==1.0.4
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.12
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
pyfiglet==0.8.post1
pygame==2.4.0
PyGetWindow==0.0.9
pyglet==1.5.26
Pygments==2.12.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMouse==1.0
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.0
PyNaCl==1.5.0
pynput==1.7.6
pyocr==0.8.3
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.0.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
PySoundFile==0.9.0.post1
pytesseract==0.3.10
pytest==8.3.2
pythainlp @ https://github.com/PyThaiNLP/pythainlp/archive/dev.zip#sha256=5f55d8cf1c6bdd35d4347a3a290467ab0b7803ebee7576b7a4c1efcb6c2d0faa
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.1.0
python-multipart==0.0.6
python-pptx==1.0.2
python-vlc==3.0.21203
pyttsx==1.1
pyttsx3==2.90
pytube==12.1.0
pytweening==1.0.4
pytz==2022.1
PyUserInput==0.1.10
pywin32==304
pywin32-ctypes==0.2.3
pywinauto==0.6.8
pyworld==0.3.3
PyYAML==6.0
pyzmq==27.0.1
qrcode==7.4.2
rapidfuzz==2.15.1
referencing==0.29.3
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.2
rfc3986==1.5.0
rpds-py==0.9.2
rsa==4.9
sacrebleu==2.3.1
safetensors==0.4.4
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.31.0
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.2
shellingham==1.5.0.post1
singledispatch==4.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.5
spacy==3.6.1
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.17
sqlparse==0.4.2
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
stats==0.1.2a0
style==1.1.0
sympy==1.10.1
symspellpy==6.7.8
tabulate==0.9.0
termcolor==2.4.0
textblob==0.18.0.post0
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tiktoken==0.8.0
tinycss2==1.4.0
tinydb==4.7.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchcrepe==0.0.20
torchvision==0.15.2
tornado==6.5.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.30.2
trio==0.30.0
trio-websocket==0.12.2
trove-classifiers==2023.7.6
typer==0.9.0
typing_extensions==4.12.2
tzdata==2022.2
uc-micro-py==1.0.2
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.23.1
v==1
virtualenv==20.24.0
wasabi==1.1.2
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.3
Werkzeug==2.3.6
whichcraft==0.6.1
wsproto==1.2.0
WTForms==3.0.1
wxPython==4.2.1
XlsxWriter==3.2.0
yarg==0.1.9
yarl==1.9.2
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2025.7.21
zipp==3.16.2
zope.event==4.5.0
zope.interface==5.4.0
